requirements.txt
# brew install poppler tesseract libmagic

unstructured[all-docs]
pillow
easyocr
lxml 
langchain
langchain-community 
langchain-groq
python_dotenv
# chromadb
tiktoken
langchain_huggingface
chromadb
streamlit
langchain_chroma
# python3 -m RAG.RAG_pipeline





llama_index-embeddings-huggingface
llama_index-llms-groq

















# llama_index 
# llama_index-embeddings-huggingface 
# llama_index-llms-groq
# # unstructured[all-docs]
# unstructured[csv, doc, pdf]
# # pillow 
# # pydantic 
# # lxml 
# # matplotlib
# # pytesseract

# # unstructured-pytesseract
# # tesseract

# torch 
# torchvision 
# torchaudio


# # brew install tesseract
# # brew install poppler
# # langchain
# # pip install 'unstructured[local-inference]'
# # cython
# # git clone https://github.com/facebookresearch/detectron2.git